Clarify and stabilize the undefined data types
We list data types that are undefined at the start of our generated
files. Stabilize the list to allow for clearer diffs, and add a better
explanation of what the data types mean.

We should expand this with a message indicating what a developer should
do to resolve the issue.